Separated ble_ota.cpp
